CMakeLists.txt: move static library test above project
In order to get CMake to only test static library linking for cross
compiler toolchains during the configure step.
CMAKE_TRY_COMPILE_TARGET_TYPE must be set above project definition. If
it is not explicitly (no checking of CMAKE_CROSSCOMPILING), the error
with the configure step testing dynmanic library linking on toolchain
with only static linking available will still fail as described in:

  https://gitlab.com/cunity/cunit/-/issues/39

This is described here:

  https://cmake.org/cmake/help/v3.10/variable/CMAKE_TRY_COMPILE_TARGET_TYPE.html

and

  https://stackoverflow.com/questions/53633705/cmake-the-c-compiler-is-not-able-to-compile-a-simple-test-program

Signed-off-by: Ryan Barnett <ryan.barnett@rockwellcollins.com>